+++ src/components/auth/index.js
@@ -1,5 +1,5 @@
 export { default as SigninForm } from './SigninForm';
 
-export { default as SignupHeader } from './SignUp/SignupHeader';
-export { default as SignupOption } from './SignUp/SignupOption';
-export { default as SignupForm } from './SignUp/SignupForm';
+export { SignupHeader } from './SignUp';
+export { SignupOption } from './SignUp';
+export { SignupForm } from './SignUp';
