+++ src/components/editProfile/EditPassword/index.js
@@ -1,0 +1,3 @@
+export { default as EditPassword } from './EditPassword';
+
+export { default as Password } from './Password';
