+++ src/components/editProfile/index.js
@@ -1,3 +1,2 @@
-export { default as EditPassword } from './EditPassword';
 export { default as DeleteUser } from './DeleteUser';
-export { default as Password } from './Password';
+export { EditPassword } from './EditPassword';
